Fix writing results of GP models to output file
Previously, the results of fitting GP models to the data were not being written to the output file correctly. This caused the output file to contain incorrect data. The code has been modified to correctly write the results to the output file, ensuring that each row contains the input file name, species name, BIC value, kernel type, and mean type for each fitted model. This fix improves the accuracy and reliability of the output file.